pyproject.toml
[tool.poetry]
name = "swipeapi"
version = "0.1.0"
description = ""
authors = ["hushchynd <hushchynd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pre-commit = "^3.3.3"
django = "^4.2.3"
djangorestframework = "^3.14.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.4"}
django-cors-headers = "^4.2.0"
easy-thumbnails = "^2.8.5"
django-cleanup = "^8.0.0"
django-allauth = "^0.54.0"
dj-rest-auth = "^4.0.1"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.7"
drf-psq = "^1.1.0"
django-filter = "^23.2"
python-dateutil = "^2.8.2"
drf-extra-fields = "^3.7.0"
celery = "^5.3.1"
redis = "^5.0.0"
drf-compound-fields = "^2.0.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
pytest-factoryboy = "^2.5.1"
model-bakery = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
